[backend] 유저 업데이트 기능 수정

+++ src/main/java/com/rubminds/api/user/service/UserService.java
@@ -83,7 +83,9 @@ public class UserService {
     }
 
     private void updateAvatar(MultipartFile file, User findUser) {
-        avatarRepository.deleteById(findUser.getAvatar().getId());
+        if(findUser.getAvatar()!=null) {
+            avatarRepository.deleteById(findUser.getAvatar().getId());
+        }
         Avatar avatar = uploadAvatar(file);
         findUser.updateAvatar(avatar);
     }
